+++ bank-feeds/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [csharp v3.0.0] bank-feeds
 ### Releases
 - [NuGet v3.0.0] https://www.nuget.org/packages/Codat.BankFeeds/3.0.0 - bank-feeds
+
+## 2023-12-04 01:29:51
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.0.1] bank-feeds
+### Releases
+- [NuGet v3.0.1] https://www.nuget.org/packages/Codat.BankFeeds/3.0.1 - bank-feeds